Conflicts:
	src/apyc.hn

+++ tests/correct/func_basic_processed.py
@@ -1,2 +1,2 @@
 def hello()::list:
-    assign::int = test
+    assign = test
